requirements_dev.txt
asn1crypto==0.24.0
attrdict==2.0.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coloredlogs==10.0
cryptography==2.6.1
cycler==0.10.0
cytoolz==0.9.0.1
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.5
docker==4.0.2
entrypoints==0.3
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.2
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.5.2
Flask==1.0.3
hexbytes==0.1.0
humanfriendly==4.18
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
jupytext==1.1.6
keeper-contracts==0.9.7
kiwisolver==1.1.0
lru-dict==1.1.6
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
ocean-cli==0.6.4
ocean-secret-store-client==0.0.1
pandas==0.24.2
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.8.1
pycryptodomex==3.8.1
Pygments==2.4.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==4.2b4
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.21.0
retrying==1.3.3
rlp==1.1.0
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
squid-py==0.6.4
terminado==0.8.2
testfixtures==6.10.0
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
web3==4.5.0
webencodings==0.5.1
websocket-client==0.56.0
websockets==5.0.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
